// packages/profile/src/generated/graphqlgen.ts
export namespace QueryResolvers {
    export const defaultResolvers = {}

    export interface ArgsGetProfile {
        username: string
    }

    export type GetProfileResolver = (
        parent: undefined,
        args: ArgsGetProfile,
        ctx: Context,
        info: GraphQLResolveInfo,
    ) => ProfileResponse | null | Promise<ProfileResponse | null>

    export interface Type {
        getProfile: (
            parent: undefined,
            args: ArgsGetProfile,
            ctx: Context,
            info: GraphQLResolveInfo,
        ) => ProfileResponse | null | Promise<ProfileResponse | null>
    }
}
// ...
    export const defaultResolvers = {}
// ...
export namespace SubscriptionResolvers {
    export const defaultResolvers = {}

    export interface ArgsFriendRequestSubscription {
        id: string
    }

    export interface ArgsFriendSubscription {
        id: string
    }

    export type FriendRequestSubscriptionResolver = {
        subscribe: (
            parent: undefined,
            args: ArgsFriendRequestSubscription,
            ctx: Context,
            info: GraphQLResolveInfo,
        ) =>
            | AsyncIterator<UserSubscriptionPayload>
            | Promise<AsyncIterator<UserSubscriptionPayload>>
        resolve?: (
            parent: undefined,
            args: ArgsFriendRequestSubscription,
            ctx: Context,
            info: GraphQLResolveInfo,
        ) => UserSubscriptionPayload | Promise<UserSubscriptionPayload>
    }

    export type FriendSubscriptionResolver = {
        subscribe: (
            parent: undefined,
            args: ArgsFriendSubscription,
            ctx: Context,
            info: GraphQLResolveInfo,
        ) =>
            | AsyncIterator<UserSubscriptionPayload>
            | Promise<AsyncIterator<UserSubscriptionPayload>>
        resolve?: (
            parent: undefined,
            args: ArgsFriendSubscription,
            ctx: Context,
            info: GraphQLResolveInfo,
        ) => UserSubscriptionPayload | Promise<UserSubscriptionPayload>
    }

    export interface Type {
        friendRequestSubscription: {
            subscribe: (
                parent: undefined,
                args: ArgsFriendRequestSubscription,
                ctx: Context,
                info: GraphQLResolveInfo,
            ) =>
                | AsyncIterator<UserSubscriptionPayload>
                | Promise<AsyncIterator<UserSubscriptionPayload>>
            resolve?: (
                parent: undefined,
                args: ArgsFriendRequestSubscription,
                ctx: Context,
                info: GraphQLResolveInfo,
            ) => UserSubscriptionPayload | Promise<UserSubscriptionPayload>
        }

        friendSubscription: {
            subscribe: (
                parent: undefined,
                args: ArgsFriendSubscription,
                ctx: Context,
                info: GraphQLResolveInfo,
            ) =>
                | AsyncIterator<UserSubscriptionPayload>
                | Promise<AsyncIterator<UserSubscriptionPayload>>
            resolve?: (
                parent: undefined,
                args: ArgsFriendSubscription,
                ctx: Context,
                info: GraphQLResolveInfo,
            ) => UserSubscriptionPayload | Promise<UserSubscriptionPayload>
        }
    }
}